Forenkle serializers, flytt mapper til test
